refactor: 카프카 에러핸들러 위치 변경 (#129)
